requirements.txt
# Core data processing libraries
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0

# Data visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Statistical analysis
scipy>=1.11.0
scikit-learn>=1.3.0

# Database connectivity
sqlalchemy>=2.0.0
pymongo>=4.5.0
psycopg2-binary>=2.9.0

# Configuration and logging
pyyaml>=6.0
jsonschema>=4.19.0

# API and HTTP requests
requests>=2.31.0

# Progress bars and CLI
tqdm>=4.65.0
click>=8.1.0

# Date/Time handling
python-dateutil>=2.8.0

# Memory efficiency
pyarrow>=12.0.0

# Web framework (for future dashboard)
flask>=2.3.0
flask-cors>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
